Update Product & Category

+++ src/main/java/com/hhd/jewelry/service/ProductService.java
@@ -10,7 +10,7 @@ public interface ProductService {
     List<Product> getProductsByCategoryName(String categoryName);
     List<Product> getProductsByCollectionName(String collectionName);
     List<Product> getProductsByCategoryNameAndCollectionName(String categoryName, String collectionName);
-    List<Product> getProductsByPriceBetween(Long minPrice, Long maxPrice);
+    List<Product> getProductsByPriceBetween(Integer minPrice, Integer maxPrice);
     List<Product> getProductsByStockQuantityGreaterThan(int stock);
     List<Product> getProductsByMaterial(String material);
     Product save(Product product);
